+++ packages/frontend/src/components/message/Message.tsx
@@ -365,6 +365,7 @@ function buildContextMenu(
             // but let's not pass `chatId` here, for future-proofing.
             msgChatId: undefined,
             highlight: true,
+            focus: true,
             msgParentId: message.id,
             scrollIntoViewArg: { block: 'center' },
           })
@@ -914,6 +915,7 @@ export const Quote = ({
         msgId: quote.messageId,
         msgChatId: undefined,
         highlight: true,
+        focus: true,
         msgParentId,
         // Often times the quoted message is already in view,
         // so let's not scroll at all if so.